arm64: morello: Enable access to capabilities in memory
This patch allows capabilities to be loaded from and stored to
memory via most mappings, specifically:
- Non-executable kernel mappings.
- All private user mappings.

Shared user mappings are explicitly excluded, to prevent
capabilities from being leaked out of the process (and therefore
address space) they belong to.

PAGE_EXECONLY is not relevant as Morello is a fixed v8.2
configuration, and therefore doesn't include EPAN. Even if it did,
a true exec-only page does not allow any read/write access so it
would not make sense to enable access to capability tags there.

Since the Morello PTE attributes are allocated in the PBHA bits, we
also need to enable them by configuring TCR_EL1 accordingly. While,
in principle, this could be done when the Morello feature is
detected, modifying the HPD/TTPBHA bits after the MMU is enabled is
potentially unsafe. Instead, we set them when TCR_EL1 is initially
configured (before the MMU is enabled).

Note that we do not need to decide whether or not to set the Morello
PTE attributes at runtime. Indeed, if Morello is not detected, the
PBHA bits will not be enabled and setting some of them has no
effect.

[Kevin: this commit no longer introduces a split between private
 and shared mappings. This is now actually introduced by a later
 commit, "arm64: morello: Explicitly add VM_*_CAPS to private user
 mappings". Making the split here became quite onerous on 6.7, and
 there is no point complicating further rebases considering the less
 invasive approach used in that later commit. This commit should
 remain functional, albeit too permissive for shared mappings.]

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>