arm64: morello: Context-switch capability GPRs and CSP
Capability registers are not separate from the base registers in
Morello. Instead, the conceptual GPRs R0 to R30 are extended to be
129-bit wide, and can be accessed as capabilities via Cn, while the
lower 64 bits remain accessible via Xn.

The most straightforward way to handle this in the kernel would be
to only store the capability GPRs (C0-C30) in pt_regs.
Unfortunately, in multiple situations, kernel code modifies specific
(64-bit) GPRs in pt_regs. It is not possible to solve this simply
using unions, as regular 64-bit stores clear the capability tag in
memory.

To avoid an invasive and potentially tricky surgery, the approach we
are taking here is to store the capability GPRs alongside the 64-bit
GPRs in pt_regs. The capability GPRs are then saved in pt_regs
during kernel_entry.

During kernel_exit, we need to reconcile the 64-bit and capability
views of the GPRs, in particular if 64-bit GPRs have been modified
since kernel_entry. This is done by "merging" each 64-bit GPR into
its capability counterpart (setting Cn's value to Xn). We only do so
if Xn is different from Cn's value, as we would otherwise
inadvertently invalidate sealed capabilities.

We use exactly the same method for CSP, which must also be
saved when entering the kernel from EL0 as the kernel itself
uses SP_EL0 to store current_thread_info(), corrupting CSP_EL0.

A side effect of this approach is that setting a 64-bit GPR in the
kernel (for instance via ptrace) does not clear the corresponding
capability GPR, it only sets its value (potentially clearing its
tag). This only makes a difference when looking at the capability
register, and is desirable in some cases.

This patch also modifies copy_thread_tls() to preserve capability
tags when copying pt_regs to the new thread.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>