arm64: Refactor user_tls accessors
There is no need for task_user_tls() to be exposed in processor.h.
Move it to process.c, and simplify it by turning into a function and
removing the #ifdef CONFIG_COMPAT, which was made unnecessary by
commit 65896545b69f ("arm64: uaccess: Fix omissions from usercopy
whitelist").

Also introduce convenience functions to save and restore the TLS
pointer in task_struct, as Morello support is going to make these
operations more complex.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>