arm64: morello: Context-switch user system registers
Morello extends a number of user-accessible (EL0) system registers
into capabilities, and adds brand new registers as well. Those
registers are unused in the kernel, so we can context-switch them
during __switch_to instead of kernel_entry/kernel_exit.

This commit adds storage for the registers in thread_struct, and:
- Initialises them to the appropriate values in
  arch_setup_new_exec().
- Context-switches them in __switch_to().
- Copies them to the new thread on clone().

ctpidr_el0 requires special handling, as tpidr_el0 is already stored
as tp_value in thread_struct and tp_value is directly accessed in
the kernel in a few places. To avoid invasive changes, we take the
same approach as for the GPRs, i.e. ctpidr_el0 is saved separately
in thread_struct, and tp_value is merged into the saved ctpidr_el0
when the user TLS pointer is restored. This means that when the
(64-bit) TLS pointer is set via ptrace or clone(), it will be merged
into the current ctpidr_el0.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>