arm64: morello: Context-switch Restricted registers
When a user thread is running in Restricted (i.e. the Executive
permission in PCC is cleared), all direct or indirect accesses to
(C)SP, (C)TPIDR and DDC are redirected to the corresponding
Restricted register (resp. RCSP, RCTPIDR and RDDC).

This is unproblematic for (R)DDC, as it does not overlap with any
register in the base Armv8 architecture, and we can therefore
context-switch RDDC just like DDC. We initialise RDDC to the null
capability, as processes are always started in Executive and the
early init code can set RDDC if necessary.

Things are unfortunately more complex for (R)CSP and (R)CTPIDR,
because they correspond to the existing SP and TPIDR registers in
the base architecture, and both of them are accessed by the kernel
for various purposes (including writing, for instance in ptrace).

To keep most operations working seamlessly when a thread is
running in Restricted, the approach taken here is to save the lower
64 bits of the Restricted registers (RSP and RTPIDR) as the thread's
SP and TPIDR values when the thread enters the kernel or is
scheduled out. Conversely, when the thread exits the kernel or is
scheduled in, the saved SP and TPIDR values are merged into the
registers that will be "active" when the thread runs again, i.e.
RCSP/RCTPIDR if the saved PCC's Executive permission is cleared, and
CSP/CTPIDR otherwise. This notably enables operations like tls_set()
to modify the register that is actually being used by the thread.

Regarding TPIDR specifically, another consequence of this approach
is that if a TLS pointer is specified on clone(), the new value will
be merged into the thread register the new thread will actually be
using (RCTPIDR if the cloning thread is running in Restricted, CTPIDR
otherwise).

Signal handling requires extra care, because signal handlers are
always run in Executive, and therefore we need to make sure that the
Executive SP is used regardless of the interrupted context. Although
signal handling does not interact with RCSP, we also save it to the
signal frame for consistency with CSP.

It is worth noting that there is only one instance in which a
thread's PCC permissions are modified by the kernel: when returning
from a signal, most registers including PCC are set to the values
saved in the signal frame on the user stack. Since the values are
entirely controlled by the user thread itself, it is up to it to
modify both the 64-bit and Restricted/Executive capability views of
all registers consistently. For instance, if the signal handler
clears the Executive permission of the PCC saved in the signal
frame, it must also update (at least) the saved SP to match the
value of RSP. TPIDR does not need special handling, as the value
stored in thread_struct is anyway out of sync while current is
running.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>