arm64: morello: Suspend support
Morello adds a number of system registers at EL1 and EL0; we need to
save and restore most of them when suspending a CPU.

All the EL0 Morello registers that are stored in thread_struct need
to be saved and restored. In particular, the capability view of
TPIDR_EL0 needs to be preserved. The remaining registers stored in
pt_regs (notably (R)CSP_EL0) will be restored during ret_to_user, so
they don't need to be handled here.

The kernel only ever writes to two EL1 Morello registers, CCTLR_EL1
and CVBAR_EL1, which we preserve in the same way.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>