arm64: morello: Kconfig entry
Everything is now in place to expose Morello in Kconfig, and enable
it by default (if the assembler supports it). There are several
caveats, detailed in the config description.

We only require assembler support for Morello at this stage. Both
Clang's integrated assembler and gas are supported, some special
care is required to ensure that both are handled correctly.

We do not support booting without VHE, so we make ARM64_MORELLO
depend on ARM64_VHE. While it should be possible to support booting
without VHE, this would be mostly pointless, since the Morello
architecture mandates the presence of VHE.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit a41f646d159054505cee3250ab35d2fc23bd4b17)
[ Christian A. Ehrhardt:
  Starting with kernel 6.18 CONFIG_VMAP_STACK is mandatory for arm64.
  This drops the Kconfig incompatibility with morello from the original
  commit. The actual issues with morello and CONFIG_VMAP_STACK must be
  addressed separately. ]
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>