arm64: morello: ptrace: Add a new sysctl to allow capability forging
Capabilities cannot currently be set via ptrace requests. We are
about to add this ability, but because it will make it possible to
inject any (valid) capability in the tracee, it is highly privileged
and needs to be enabled explicitly.

This commits adds a new "cheri.ptrace_forge_cap" sysctl that
controls this ability. It is set to 0 by default, and must be set to
1 to enable setting capabilities via ptrace (globally).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit dcfadb2e62ab15ed0415454adb5a69f0129af5ac)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>