arm64: morello: Represent capabilities using uintcap_t directly
The cap128_t type was introduced at a time when the kernel was built
with a Morello-unaware compiler, providing an opaque type that could
only be manipulated in assembly. Now that we have enabled Morello
support in C, it does not make sense to keep this abstraction; just
represent capabilities using the compiler-provided type.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>