arm64: memory: Always return a u64 in untagged_addr()
The value returned by untagged_addr() should not be dereferenced, so
it does not make sense to return the same type as the input if it is
a pointer. Always return a u64 instead.

To make it clear that a __user pointer may be passed, also add an
intermediate cast to user_uintptr_t.

arch_untagged_si_addr() expects untagged_addr() to return a pointer,
so we need to fix it up right now. The reason why it is given and
returns a pointer is that struct siginfo's si_addr is represented as
a pointer; but it is never a valid capability, only a raw 64-bit
address. as_user_ptr() is therefore the conversion function we want
here.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit 08dc9ef12dc1af8cbad83b3746f8085ace71a974)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>