linux/types.h: Introduce ptraddr_t
Introduce the new standard type that explicitly represents an
address, ptraddr_t. In the pure-capability ABI, the size of
addresses remains unchanged, unlike pointers.

If the compiler does not define the appropriate macro, fall back to
`unsigned long`, which is the type currently used throughout the
kernel to represent addresses.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>