linux/user_ptr.h: Introduce user_ptr conversion helpers
On certain architectures, in particular those implementing the
pure-capability uABI, the conversion from an integer to a user
pointer needs to be implemented in a different way depending on the
nature of the input integer / address.

To allow for this, introduce new *_to_user_ptr* helpers that
architectures can override, as well as a new as_user_ptr() helper
for cases where the resulting user pointer is not expected to be
dereferenceable (useful for instance when storing an arbitrary
integer in a user pointer). as_user_ptr_strict() additionally checks
that the input type is as specified.

Conversely, a new user_ptr_addr() helper is added to extract the
address of a user pointer using the appropriate casts.

u64_to_user_ptr() is reimplemented in terms of as_user_ptr_strict().
Its usage is deprecated, as the semantics of *_to_user_ptr* helpers
is now to produce dereferenceable user pointers, unlike
u64_to_user_ptr().

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>