ipc,mm: Fix user pointer downcasts
Do not cast a user pointer directly to a type that cannot hold a
CHERI capability, use user_ptr_addr() to extract its address.

The downcasts in shmat/shmdt will require more attention for full
PCuABI support as additional capability checks will be required
there. Conversely the downcasts in process_*() syscalls are not
annotated because they concern pointers in remote processes, which
are not subject to capability checks.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>