lib: Fix user pointer downcasts
Do not cast a user pointer directly to a type that cannot hold a
CHERI capability, use user_ptr_addr() to extract its address.

Most of these downcasts are used to obtain the alignment of the
pointer's address, and are therefore harmless. Some are actually
used to obtaining the address of the user mapping in order to access
the memory and will require additional capability checks to fully
support PCuABI, TODOs are added accordingly.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>