fs/signalfd: Fix user pointer downcasts
Do not cast a user pointer directly to a type that cannot hold a
CHERI capability, use user_ptr_addr() to extract its address.

ssi_ptr is currently a __u64 so there is no problem with simply
assigning ssi_ptr to the address of the input si_ptr from a uapi
perspective. It may be desirable to preserve the full user-provided
capability in PCuABI, however this would require not only modifying
the uapi (struct signalfd_siginfo) but also copying capability tags
to userspace in syscalls such as read(), which is non-trivial.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>