fs/(autofs,ext2, ext4): Use proper typecast for capability type
With the introduction of capabilities and PCuABI being enabled,
the ioctl handler, when dealing with the user pointers does expect
a capability. Furthermore compat_ptr, for related platforms,
will/should provide one so in order not to mess with that, use proper
casting. This also addresses a compilation warning triggered by otherwise
implicit conversion that might lead to unexpected behaviour when operating
on capabilities.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>