kselftests/arm64: Change the running stack to a temporary one
The kernel provided stack needs to be tested before it can be safely used.
Some C programs rely on the stack to perform basic tasks, so not using it
is not an option. To avoid testing the stack in assembly we install a
temporary one that can be used to perfom initial tests. For tests which do
not care or rely on the kernel one we provide a subroutine to restore it.
There is also a switch to disable the functionality altogether if it's
undesired.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>