scsi: use ptr_to_compat() to convert compat pointers
For compat tasks, put_sg_io_hdr() and scsi_put_cdrom_generic_arg()
convert the command header struct to the corresponding compat struct,
before copying it to userspace. When converting a __user pointer to a
compat_uptr_t, a cast to uintptr_t is currently used. With the Arm
Morello compiler, this produces compiler warnings of the form:

    drivers/scsi/scsi_ioctl.c:641:16: warning: the following conversion will result in a CToPtr operation; the behaviour of CToPtr can  be confusing since using CToPtr on an untagged capability will give 0 instead of the  integer value and should therefore be explicitly annotated
[-Wcheri-pointer-conversion]
                            .dxferp          = (uintptr_t)hdr->dxferp,
                                               ^~~~~~~~~~~~~~~~~~~~~~

To extract an address from a CHERI capability, it needs to first be
converted to __uintcap_t type and only then smaller integer types.
Fortunately there is already a helper function to perform this operation
for compat pointers, which also preserves existing casts on non-CHERI
kernels: ptr_to_compat(). Use it here. There should be no functional
change on non-CHERI kernels.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>