net: fix user pointer downcasts in compat ioctl handlers
Do not cast a __user pointer to an unsigned long, as it may be a CHERI
capability. Instead cast it to user_uintptr_t, in order to preserve the
capability. This fixes compiler warnings of the form:

net/unix/af_unix.c:2642:31: warning: the following conversion will result in a CToPtr operation; the behaviour of CToPtr can  be confusing since using CToPtr on an untagged capability will give 0 instead of the  integer value and should therefore be explicitly annotated [-Wcheri-pointer-conversion]
        return unix_ioctl(sock, cmd, (unsigned long)compat_ptr(arg));
                                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Note that unix_ioctl still takes an unsigned long as an argument. It
will need to be fixed later along with the rest of the sockets
interface. This patch just fixes the few places that the compiler warns
about when CONFIG_COMPAT is enabled.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>