tty: fix user pointer downcasts in compat ioctl handlers
Do not cast a __user pointer to an unsigned long, as it may be a CHERI
capability. Instead cast it to user_uintptr_t, in order to preserve the
capability. This fixes compiler warnings of the form:

drivers/tty/pty.c:500:33: warning: the following conversion will result in a CToPtr operation; the behaviour of CToPtr can  be confusing since using CToPtr on an untagged capability will give 0 instead of the  integer value and should therefore be explicitly annotated [-Wcheri-pointer-conversion]
        return pty_bsd_ioctl(tty, cmd, (unsigned long)compat_ptr(arg));
                                       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Note that the called functions (vt_ioctl, pty_bsd_ioctl,
pty_unix98_ioctl) still take an unsigned long as an argument. They will
need to be fixed later along with the rest of the tty interface. This
patch just fixes the few places that the compiler warns about when
CONFIG_COMPAT is enabled.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>