vt_ioctl: use ptr_to_compat() in compat_kdfontop_ioctl()
The KDFONTOP compat ioctl handler converts between a struct
console_font_op and a struct compat_console_font_op. When converting a
compat user pointer to a native user pointer it uses the appropriate
compat_ptr() function. However, when converting in the other direction
it simply uses a cast to unsigned long. With the Arm Morello compiler,
this produces compiler warnings of the form:

    drivers/tty/vt/vt_ioctl.c:1169:48: warning: the following conversion will result in a CToPtr operation; the behaviour of CToPtr can  be confusing since using CToPtr on an untagged capability will give 0 instead of the  integer value and should therefore be explicitly an
notated [-Wcheri-pointer-conversion]
            ((struct compat_console_font_op *)op)->data = (unsigned long)op->data;
                                                          ^~~~~~~~~~~~~~~~~~~~~~~

To extract an address from a CHERI capability, it needs to first be
converted to __uintcap_t type and only then smaller integer types.
Fortunately there is already a helper function to perform this operation
for compat pointers, which also preserves existing casts on non-CHERI
kernels: ptr_to_compat(). Use it here. There should be no functional
change on non-CHERI kernels.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>