autofs: do not use compat_ulong_t to define AUTOFS_IOC_SETTIMEOUT32
The AUTOFS_IOC_SETTIMEOUT32 ioctl command is defined using
compat_ulong_t as the data type. As part of the work to enable a 64-bit
compat layer, compat_ulong_t will soon be changed to be 64-bit, which
will cause AUTOFS_IOC_SETTIMEOUT and AUTOFS_IOC_SETTIMEOUT32 to have the
same values, triggering an error in the switch statement in
autofs_root_ioctl_unlocked().

To fix this, use another 32-bit type in AUTOFS_IOC_SETTIMEOUT32.

It's also worth noting that compat_ulong_t is an in-kernel type and is
not defined in the uapi headers, so shouldn't be used there at all.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>