block: fix user pointer downcasts in compat ioctl handlers
Do not cast a __user pointer to an unsigned long, as it may be a CHERI
capability. Instead cast it to user_uintptr_t, in order to preserve the
capability. This fixes compiler warnings of the form:

drivers/mmc/core/block.c:786:40: warning: the following conversion will result in a CToPtr operation; the behaviour of CToPtr can  be confusing since using CToPtr on an untagged capability will give 0 instead of the  integer value and should therefore be explicitly annotated [-Wcheri-pointer-conversion]
            return mmc_blk_ioctl(bdev, mode, cmd, (unsigned long) compat_ptr(arg));
                                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>