scsi: sg: fix type of pointer field in compat_sg_io_hdr
The dxferp field in struct compat_sg_io_hdr contains a compat pointer,
but it is currently defined as a compat integer. This is going to break
Arm Morello systems with 64-bit compat where a compat pointer will be
defined to be 64-bit but a compat integer will still be 32-bit. To make
it work, change the type of the field to be compat_uptr_t, which is of
the right compat pointer size on all types of systems.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>
(cherry picked from commit daf13b73370f2e6d787cef84428cc255c3f41ec2)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>