drivers/android/binder: Create valid user pointers
In certain cases, Binder needs to obtain user pointers from
user-specified structs. Unfortunately, at the moment these structs
represent all pointers using the binder_uintptr_t type, which is
always an integer. As a result, in PCuABI, we need to create valid
user pointers from the user-provided addresses.

Note: this is a minimal fix that preserves the existing uapi. A
proper port of Binder for PCuABI should modify the uapi to operate
on (capability) user pointers directly, without the need to
synthesise capabilities on the kernel side.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>