arm64: signal: Keep AArch32 signal specific code within COMPAT32
The compat layer for PCuABI is compat64 and hence AArch32 specific
code (signal32.c and sigreturn32.c) are compiled only when COMPAT32
config is present.

The native sigreturn syscall handler is temporarily used for compat
sigreturn syscall to let the compilation go through. This will be
removed after the full compat64 signal support is added.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>