fs/compat_binfmt_elf: Modify compat layer to support LP64 interface
Now most of the building blocks are present for the compat layer to
support LP64 interface, so redefine few of the macros to be picked by
compat_binfmt_elf.c correctly.

Few macros redefined are common for both 32 and 64 bit compat and were
added for PCuABI support in native elf binary loader.

Re-use ARCH_DLINFO and elf_check_arch redefinitions for 64-bit compat.

Also define a macro COMPAT_ARCH_DLINFO for AArch64 in case of PCuABI.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>