arm64: morello: Fix CTPIDR handling in ptrace
Since commit "arm64: morello: Add support for TLS as capabilities",
thread.uw.tp_value is a capability in PCuABI. A few places have not
been updated accordingly, in particular
morello_flush_cap_regs_to_64_regs(), which truncates CTPIDR to a
64-bit integer before storing it in tp_value. Cast it to
user_uintptr_t instead to make it work as intended with and without
PCuABI. This fixes ptrace(PTRACE_SETREGSET, NT_ARM_MORELLO).

Conversely, tls_get() still operates on 64-bit integers (to preserve
the existing uapi), but the size it writes to user memory is
currently inferred from the size of tp_value, which is a full
capability in PCuABI. Cast it to unsigned long explicitly to avoid
this. This fixes ptrace(PTRACE_GETREGSET, NT_ARM_TLS).

While at it, also add casts in tls_set() to make the conversions
explicit. This change is only cosmetic.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>