efi/arm64: morello: Keep EFI stub in pure aarch64 mode
With Morello, EFI stub ends up in a rather grey area, with no safety net
in place to make sure it does not rely on anything Morello specific:
for one, there is no need for it to, and actually it should not as there
is no guarantee for context switching to be considered safe in this case.
Make the EFI stub completely unaware of Morello support by disabling any
Morello (and related) options.

This is far from being perfect but in the end, it gets the job done.

Signed-off-by: Beata Michalska <beata.michalska@arm.com>