arm64: syscall: Don't modify C0 in compat64 syscall
In compat64, the syscall never returns a valid capability. Therefore
setting C0 to the return of the syscall will always clear the metadata
of the register. This is problematic in the case of sigreturn syscall,
which must restore all registers, including C0. The change aligns
compat64 with the standard ABI (!PCuABI), where X0 is always merged into
C0 without it being cleared, regardless of the syscall.

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>