linux/compat.h: Fix size of compat_siginfo_t in compat64
From a uapi perspective, siginfo_t has a fixed size of 128 bytes
regardless of the ABI and architecture. This is achieved through a
union that includes a padding array. In compat_siginfo_t, the union
does not include the "header" of common members (si_signo / si_errno
/ si_code), as a result the padding array needs to account for them.

The calculation is unfortunately invalid in compat64 as it assumes
that the _sifields union has the same alignment as an int (4), which
is no longer true as compat_uptr_t is 8 bytes, requiring the whole
union to be aligned to 8 bytes. Fix this by accounting for the
alignment requirement in the calculation.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>