arm64: morello: Disable v8.3+ AS-dependent features
When Morello support is enabled, -march=morello is passed to compile
and assemble all files. This does not play well with v8.3+ features
that rely on assembler support, in which case a different -march is
passed to the assembler. The ARM64_ASM_ARCH macro (expanding to
.arch ...) also gets used for inline assembly in C, overriding
-march=morello.

Morello being a fixed v8.2 configuration (without runtime
detection), there is in fact no point in supporting these v8.3+
features when Morello support is enabled. Unfortunately, the
assembler flags get enabled by default if the assembler supports it,
so we need to introduce a strange dependency on !ARM64_MORELLO for
the problematic AS_HAS_* options to prevent the issue.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit 1ccb742338ba76294379eb84874af83392519c19)
[ Christian A. Ehrhardt:
  Upstream dropped support for older gcc/clang versions and seems to
  assume that the compiler supports at least armv8.4-a.
  FIXCHERI: Maybe set asm-arch to armv8.2-a for morello?
]
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>