fs/binfmt_elf: Fix auxv parsing for compat ELF binaries in /proc/pid/auxv
For a compat ELF binary, /proc/pid/auxv will output an extra AT_NULL
entry or garbage at the end. This is because it interprets the auxv
elements using their native layout. The patch moves the auxv parsing
to the binfmt handler, where the correct layout of the auxv elements
is used. This patch is inspired by: https://lkml.org/lkml/2006/4/10/20

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>