iov_iter: Handle struct iovec and struct kvec separately
There is an underlying assumption that struct iovec and struct kvec have
identical layouts. This assumption doesn't hold in the PCuABI, so handle
the two structs separately. Note that iterate_iovec macro is unchanged
and covers both kvec and iovec because it only assumes the names of the
struct members are the same, not their types. Therefore, it still
holds for PCuABI.

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>