pps: Update compat ioctl handler for compat64
Update pps_cdev_compat_ioctl() to perform the appropriate
compat / native user pointer conversion using compat_ptr(), and pass
on the result to pps_cdev_ioctl().

The special handling for PPS_FETCH in pps_cdev_compat_ioctl() is not
actually needed for compat64, and in fact does not work as it relies
on packed structs that result in the compat_* structs diverging from
the native structs (even in compat64). Fix that by only using this
code in compat32.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>