uaccess: Add pointer-preserving routines
This patch introduces new *_with_ptr() routines that guarantee that
(user) pointers contained within the transferred data are fully
preserved.

The main motivation for adding these new routines is to provide a
generic interface that is compatible with the pure-capability uABI,
where user pointers are tagged and are not preserved by ordinary
uaccess routines.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>