epoll: Add compat handling of epoll_event
A subsequent patch is going to change struct epoll_event to enable it to
support new architectures. On such architectures, the current struct
layout still needs to be supported for compat tasks. In preparation for
that change, add a struct compat_epoll_event and handle it in epoll
syscalls.

Note, it may be cleaner to have a separate compat entry point for
epoll_ctl. For now, use only the native entry point to avoid changing
the syscall table on all architectures.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>