fs/aio: Use user pointer type for iocb.aio_buf
The 'aio_buf' member of struct iocb may contain a user pointer. On
some architectures (for example CHERI) a user pointer is a 129-bit
capability, so the __u64 type is not big enough to hold it. Use the
__kernel_uintptr_t type instead, which is big enough on the affected
architectures while remaining 64-bit on others.

In addition, special copy routines need to be used when copying user
pointers from userspace. Use these for iocb.aio_buf.

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>