fs/binfmt_elf: Add additional Aux vector entries
To aid in libc development, additional auxiliary vector entries are added
in the native elf loader. This is supplemented by code that determines
the beginning and end addresses of the load segments in the elf and
interpreter. A full description of the entries and their constant values
can be found in the PCuABI specification at [1].

[1]: https://git.morello-project.org/morello/kernel/linux/-/wikis/Morello-pure-capability-kernel-user-Linux-ABI-specification

Signed-off-by: Sherwin da Cruz <sherwin.dacruz@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>