elf: Return result from START_THREAD macro
The START_THREAD arch hook called in binfmt_elf is modified to return a
value which the elf loader returns as the final result. Since no arch
has specialised this macro it is being modified to allow any arch
specialising this macro to either 1. Abort execution of an elf binary
from within start_thread() by returning a negative value or 2. Place a
positive value in the first result register for the executing program to
access by returning the positive value.

This modified macro also takes in the bprm struct instead of the sp to
allow any arch to access more information about the binary contained
within. Compatibility is maintained in previous versions by passing the
sp from bprm to the default start_thread() implementation.

This required changing kernel_execve() to always return 0 on success
instead of the aforementioned positive value. This is because only a
return value of 0 would indicate that the init process has started
correctly during boot, in which kernel_execve() is called.

Signed-off-by: Sherwin da Cruz <sherwin.dacruz@arm.com>