mm/memory: Create appropriate user pointer for uaccess
__wp_page_copy_user() sometimes needs to read a page directly via a
user mapping; to perform the uaccess we therefore need to create a
valid user pointer from the address of the page.

uaddr_to_user_ptr_safe() should no longer be used for that purpose.
Instead, we use make_user_ptr_for_read_uaccess() to create a user
pointer with appropriate bounds and permissions (in PCuABI).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>