tracing/syscalls: Allow amending metadata macro arguments
The way in which the tracing framework defines its syscall metadata is
tightly coupled with the syscall wrappers and those might be subject to
architecture specific fine-tuning. For most cases this is not an issue,
as SYSCALL_METADATA is simply piggy-backing on the arguments provided to
those wrappers, though for arm64/Morello this proves to be problematic.
To overcome this, split the SYSCALL_METADATA into a double-phased setup,
allowing amending the arguments themselves, so that whatever changes the arch
code decides to do there, get sanitised prior to expanding the actual metadata
wrapper. Additionally, the order of the arguments here gets align with
the rest of the syscall framework.

Tested-by: Zachary Leaf <zachary.leaf@arm.com>
Signed-off-by: Beata Michalska <beata.michalska@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>