arm64: morello: Explicitly add VM_*_CAPS to private user mappings
Don't add PTE_*_CAPS to default user mappings in pgtable-prot.h.
Instead, add VM_{READ,WRITE}_CAPS to all private user mappings.
This ensures that the rc/wc smaps flags are set.
The instances where the VM_*_CAPS flags are added by default are:
  - standard mappings created by the kernel itself
  - the stack and brk area by updating the VM_DATA_DEFAULT_FLAGS flags
  - private user mappings created by userspace through mmap()

This is a partial revert of:
  "arm64: morello: Enable access to capabilities in memory"
(reverting the addition of the PTE flags to the private user mappings).

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>