arm64: compat: Use appropriate root cap in compat_ptr() in PCuABI
All capabilities in hybrid (here compat64) should now be derived
from cheri_user_root_allperms_cap. There is no helper to derive
capabilities from this root, so do it manually. This is specific to
PCuABI, but the generic implementation of compat_ptr() is
appropriate in !PCuABI, so we can simply #ifdef the whole
definition.

Also update the comment as this does not pertain to PCuABI as such.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>