net/compat: Don't pack structs in COMPAT64
The __packed attribute for struct compat_group_req led to a failure of
the LTP test accept02 when trying to join a multicast group.
Keep it for COMPAT32 but remove it in COMPAT64 by introducing a new macro:
NET_COMPAT_PACKED. Do the same for the other structs used here.

Signed-off-by: Teo Couprie Diaz <teo.coupriediaz@arm.com>