arm64: morello: Build arbitrary user caps using appropriate root
As a privileged operation (disabled by default), arbitrary
capabilities may be created in a process via ptrace. Now that we
have defined generic user root capabilities, make use of
cheri_user_root_allperms_cap (the "root of roots") to build these
capabilities from, instead of the kernel root capability. Change the
name of the helper accordingly and clarify that it should not be
used for unprivileged operations.

Using cheri_user_root_allperms_cap is a clear improvement, however
this still allows creating capabilities that could not otherwise be
obtained in PCuABI (e.g. with Seal combined with other permissions).
Considering the additional complexity that restricting this further
would entail, and the operation being only involved in privileged
debugging, this situation is considered acceptable.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>