uapi/socket: align with long rather than void*
void* is used to align to 32-bits or 64-bits on 32 and 64 bits archs.
The pure-capability ABI breaks this assumption as the pointer size is
different than the word size, which is where the alignment requirement
comes from.
Indeed, the pointer size, and so void *, is 16 bytes in purecap while
the expected alignment is still 8 bytes, as it is a different ABI but
still for a 64-bit arch.

Change it to an unsigned long. Musl uses the same type for its definition
of struct sockaddr_storage.

Signed-off-by: Teo Couprie Diaz <teo.coupriediaz@arm.com>