kselftests/arm64: morello: Define uintcap_t for GCC
Clang toolchain defines uintcap_t in stdint.h present in the system
include path obtained with -print-file-name=include option.

In case of GCC toolchain uintcap_t, it is not present in the system
include path but in the libc include path. As the kselftests does
not use libc at the moment so define uintcap_t.

Signed-off-by: Amit Daniel Kachhap <amit.kachhap@arm.com>