linux/user_ptr.h: clear up as_user_ptr() usage
as_user_ptr() is intended to be used where an arbitrary integer e.g. an
error code is stored in a __user ptr.

The current implementation can be somewhat confusing in that it looks
like it can be used as direct replacement for u64_to_user_ptr e.g. in
PCuABI, where u64 addresses in the kernel-user interface are being
replaced with capability containing types such as __kernel_uintptr_t.

Currently, passing a valid capability e.g. a __kernel_uintptr_t,
__uintcap_t or user_uintptr_t etc to as_user_ptr() will result in a cast
to (void __user *) and a valid capability/pointer that can be
dereferenced. This is contrary to the code comment and intended usage
for as_user_ptr().

Add a cast to (u64) before the cast to (void __user *)(user_uintptr_t)
to make clearer the intended usage. This also always results in a null
capability that cannot be dereferenced.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>
(cherry picked from commit 1e06da936b0389360beb8b0552fe306d77360456)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>