io_uring: Implement compat versions of uAPI structs and handle them
Introduce compat versions of the structs exposed in the uAPI headers that
might contain pointers as a member. Also, implement functions that convert
the compat versions to the native versions of the struct.

A subsequent patch is going to change the io_uring structs to enable them
to support new architectures. On such architectures, the current struct
layout still needs to be supported for compat tasks.

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>