ptp: Modify ptp_ioctl to use user_uintptr_t
The third argument of ioctl handlers is a user input which can
contain either a pointer to data in the user space or any number.
This commit changes the type of the argument from unsigned long
to user_uintptr_t so that we can accommodate the bigger size of
capabilities and accept them from user space correctly without
discarding their metadata.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>