fs: Fix user_uinptr_t behaviour with ioctl
The third argument of ioctl handlers is a user input which can
contain either a pointer to data in the user space or any number.
This commit updates the signature of ext4_ioctl_checkpoint, which
is a child handler of __ext4_ioctl. __ext4_ioctl was updated to
correctly accept and supply a user_uintptr_t but the signature of
ext4_ioctl_checkpoint was not updated to reflect this change.
Moreover the value of the argument is casted to unsigned long when
outputted to the debug log to match the expected type.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>