net: bridge: Modify br_dev_read_uargs to accept user_uintptr_t
The function br_dev_read_uargs copies a list of arguments from
userspace. Normally, they may contain pointers or real numbers.
When working with capabilities, the type unsigned long used for
this purpose is not big enough to hold them.
This commit changes the type of args from unsigned long to
user_uintptr_t for this purpose. As a consequence the type of the
variables fed to this function is also updated, and casted to
unsigned long when treated as a number.
It also fixes the type of argp, as it points to an entry of args,
which lives in the kernel space, and not in the user space as the
type suggests.

Fixes: ("net: Fix the position of __capability in double pointer")
Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>