arm64: signal: Avoid unnecessary capability arithmetic
sp_top is only used to compute the size to provide to access_ok().
It should therefore be just an address, not a capability (in
PCuABI). GCC rightly warns about this situation, as subtracting two
capabilities is ambiguous (it may be interpreted as a capability
operation subtracting the address of the LHS with the address of the
RHS, while here we expect a simple integer subtraction).

Fixes: ("linux/sched/signal.h: Modify the stack pointer to user_uintptr_t")
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>