compat: Fix shift warning in compat64
Commit "compat: make compat_get_bitmap work with 64-bit
compat longs" made compat_{get,put}_bitmap() compat64-friendly by
adding a separate path when compat_long_t is the same size as long.

Even though this condition can be evaluated at compile time and the
inactive path eliminated completely, the compiler still parses both
paths. As a result, in compat64, GCC complains about the shift in
the 32-bit path, because BITS_PER_COMPAT_LONG is 64 in this case and
it does not make sense to shift a (64-bit) integer by more than 63.

Avoid this warning by shifting by BITS_PER_LONG / 2 instead, which
is equivalent as we already assert above that compat_long_t is half
the size of long in this path.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>