fcntl: Fix pointer conversion in compat
The compat fcntl handler currently does not perform simple pointer
conversions for all of the commands which take a user pointer as
an argument.

This commit adds this new case to the switch in the handler, by
converting the argument before calling do_fcntl for the relevant
commands.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>
(cherry picked from commit 631fb85951dd4abcfcc2dc8cee5ccdc0ffb1658a)
[ Christian A. Ehrhardt: Use fd_file(f) instead of f.file ]
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>