net: socket: Fix compat ioctl handler
The compat_sock_ioctl_trans handler currently passes the argument from a
compat syscall directly to the generic ioctl handler without changing
it. With PCuABI the ioctl handler can either take a capability or a real
number value. Given that the compat handler takes a pointer or a number,
this needs to be converted correctly when passed along.

This commits introduces changes that pass the argument as a capability
where required.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>