aio: Use user pointer type in the io_event struct
The libaio library may set up I/O completion callback functions for I/O
requests. The callback function pointer is stored in the data field of
the io_event struct, and the pointer to the original iocb struct is
stored in the obj field. Therefore, both the data and obj fields of the
io_event struct might hold user pointers.

In the PCuABI, a user pointer is a 129-bit capability, so the __u64 type
is not big enough to hold it. Use the __kernel_uintptr_t type instead,
which is big enough on the affected architectures while remaining 64-bit
on others. Additionally, use the special copy routine when copying user
pointers to userspace.

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>