kselftests/arm64: morello: mmap: Add capability range testcases
Overwriting part of a mapping within an existing reservation is allowed
with the use of the MAP_FIXED flag. Whereas any attempt to write beyond
the bounds of the existing capability would cause the syscall to fail.

Address space management syscalls that manipulate a given mapping are
restricted to the range owned by the capability. Any attempt to manage
mappings beyond this range will result in failure of the syscall.

Signed-off-by: Chaitanya S Prakash <chaitanyas.prakash@arm.com>
(cherry picked from commit 9231fa2823617ca458f88353e648152192439eae)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>