kselftests/arm64: morello: mmap: Add permission check testcases
When a capability is created it is assigned the maximum permissions it
may ever have by passing PROT_MAX(max_prot) as one of the prot flags.
Any attempt to increase the permissions beyond this would result in
failure of the syscall. An owning capability returned by mmap() does
not include LoadCap and StoreCap permissions if the mapping is shared.
A check to verify the same has been added.

mremap() doesn't take a prot argument and retains the permissions of the
original capability. If the permissions of the new capability exceed the
the set of permissions of the old capability, the syscall fails with a
-EINVAL error code. Tests to verify the above behaviour have been added.

Signed-off-by: Chaitanya S Prakash <chaitanyas.prakash@arm.com>
(cherry picked from commit fea3c7128ba3c29df74ae48f5f5020628bbf6dc5)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>