fs: Handle exact bounds for argv and envp
argv and envp strings can be large enough so that a capability
cannot be created to access them exclusively: their bounds would
not be exactly representable and go outside of the strings.

During allocation on the stack, align the strings that need an adjusted
base and pad if needed for alignment or a representable length.
Handle this potential padding in fs/binfmt_elf by detecting extra 0s.

This is done irrespective of the calling binary or loaded binary ABIs.
Indeed, at the point where strings are put on the stack during `execve`
we can only know the ABI of the calling binary, but not that of the loaded
binary.
Thus the compat path will be slightly slower and the memory used by the
strings will be larger, as if to create capabilities with exact bounds.
However, this should only happen in rare cases where argv or envp strings
are more than a couple thousand characters.
Even when it does, the memory impact should stay minimal.

Add a helper function to put already created capabilities on the stack
during elf setup.

Additionally, align the newly created stack to a multiple of the
maximum argument length to make sure the relocated arguments will still be
representable with their new location in a PCuABI kernel.

Signed-off-by: Teo Couprie Diaz <teo.coupriediaz@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>