gup: Add explicit capability checks
With PCuABI we want to make sure that any userspace capability is
correctly validated.

In this commit the fault_in_safe_writeable function is attempting
to fault in the pointed user page for writing purposes.
In order to do so it uses fixup_user_fault, which considers the
address and ignores any capability metadata. Therefore, it is
necessary to perform an explicit capability check to make sure
that the provided pointer has the appropriate bounds and
permissions to write to the specified address range.

This also requires to change the signature of the function by
dropping the const qualifier to the pointer argument, as
check_user_ptr_write() expects a non-const pointer. This is
acceptable given that the expectation is for the pointed region
of memory to be writeable.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>
(cherry picked from commit e47fca9f0f316ef25b0ced7a25faa08d06b2a87b)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>