usb: core: Add explicit capability checks
When working in PCuABI, userspace pointers are capabilities that carry
metadata such as bounds and read/write permissions.

In the scenario of the userspace submitting a USB Request Block, data
transfer between host and device may be handled via a kernel mapping
instead of the referenced user mapping, potentially using DMA.
As a consequence, in that case, the memory accesses are not
performed via uaccess and the capability to the user buffer used to
submit the URB is not checked.

This commit adds an explicit capability check for transfers via
a usb_memory region, to ensure that userspace has the right
permissions and bounds to submit the URB.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>