futex: Add explicit capability checks
When working with PCuABI, a user pointer is a capability carrying
metadata such as access bounds and permissions. These are automatically
enforced in hardware when a direct access is performed, which triggers a
fault if either the bounds or permissions checks failed.

Whenever a futex key is accessed and an access fault occurs, the
assumption is that the page containing the futex key must have faulted
and the code proceeds to fix it up. Because no distinction is made
between a capability fault and a page fault, an explicit capability
check needs to be employed so we can fail early and avoid entering
an infinite loop attempting to fixup a page fault.

Also, while get_futex_key looks like a valid candidate for an explicit
capability check, that is not the case as the user space pointer is
never dereferenced as it is instead directly accessed by the futex
operations later on.

Signed-off-by: Luca Vizzarro <Luca.Vizzarro@arm.com>
(cherry picked from commit c80f530ae19267b737e7414696f09a7b685dddb7)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>