aio: Change aio_context_t to a user pointer
The aio_context_t represents a token that allows the userspace to query
a specific aio context. It can also be used as a pointer to the
userspace mapping of aio_ring buffer. In some userspace software,
including libaio, it's used directly to access the aio_ring. Therefore,
aio_context_t must be able to hold a full capability. As
__kernel_ulong_t type is not large enough to fit a user pointer in the
PCuABI, change the aio_context_t type to void __user *.

As aio_context_t is still mainly used as a token id, it should not be
modified in any way by userspace. Therefore, make a full capability
comparison in lookup_ioctx() to match the aio_context_t.

Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>