arm64: uaccess: Extract user address before untagging
untagged_addr() is about to be modified to work on addresses only.
Accordingly, modify access_ok() so that it extract the user
pointer's address before passing it to untagged_addr().

We also need to create a "fake" user pointer to pass to
__access_ok(). This is unproblematic as __access_ok() only checks
the address itself; its prototype is left unchanged to avoid wider
disruption.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>