kselftests/arm64: morello: mmap: Add mremap() reservation check testcases
Attempting to remap a range larger than what is owned by the capability
triggers a -EINVAL error. Additionally, mappings that have to be moved
in order to satisfy the new constraints, expect the MREMAP_MAYMOVE flag
to be specified. Failure to do so triggers the -ENOMEM error. Within the
bounds of a reservation, mremap() can be used to grow the mappings
in-place. Tests to verify this behaviour have been added.

Signed-off-by: Chaitanya S Prakash <chaitanyas.prakash@arm.com>