kselftests/arm64: morello: mmap: Add brk() testcase
As the mechanism of brk() depends on implicit address space reservation
by moving the program break, it is unfavourable to the capability model.
Hence an assumption is made that brk() is unnecessary and allocators
making use of it can use mmap() instead. If used, it returns -ENOSYS.
A test to verify this behaviour has been added.

Signed-off-by: Chaitanya S Prakash <chaitanyas.prakash@arm.com>
(cherry picked from commit fa65c5dcb15761edb3fcdcc999735a5df8016f88)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>