arm64: vdso: Build a pure-capability vDSO
Modify the vDSO Makefile such that it can also build a vDSO for the
purecap ABI and change the fallbacks to use capability registers
when running in purecap. Therefore, under PCuABI, vDSOs will be built
for both aarch64 (COMPAT64) and a purecap.

Note that this change only implements the build of the purecap vDSO.
Mapping of this vDSO by the kernel into user address space will be done in a
subsequent patch.

Signed-off-by: Aditya Deshpande <aditya.deshpande@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>