arm64: morello: Disable KSM in defconfig
KSM (Kernel Samepage Merging) considers that two pages can be merged
if their contents are identical. Morello makes it harder to tell
whether this is truly the case as tags need to be compared, in
addition to data. Tag access is currently enabled for all private
mappings, which means that the extra tag comparison would be
required in most cases. Given that it is unclear whether KSM would
still be worthwhile with such additional overhead, let's disable it
for now.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>