mm/mremap: Add PCuABI capability handling in mremap
Use the recently introduced PCuABI reservation interfaces to add
different parameter constraints for mremap syscall. The capability
returned by the syscall is either the same as the input capability or
a new one if a new reservation is created, with the same permissions
as the pointer to the original mapping, as per the PCuABI
specification.

Signed-off-by: Amit Daniel Kachhap <amitdaniel.kachhap@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>