bpf: compat64: add handler and convert bpf_attr in
The bpf syscall does not require a compat handler for 32-bit compat.
This is achieved by using u64 instead of pointer types in the bpf_attr
union used to pass arguments to the syscall. This means that in a system
where pointers are 32-bit, the struct/union layouts and offsets are the
same as in a 64-bit arch, since the u64 field is split into two u32
fields/registers.

This greatly simplifies 32-bit compat at the small cost of requiring
casting pointers passed in through the uAPI to u64 (generally via
ptr_to_u64() helper functions).

This poses a problem in architectures where user pointers are longer
than 64b such as Morello/PCuABI where pointers are represented as 129b
capabilities. In order to extend the bpf syscall interface to accept
capabilities and still retain compatibility with the existing 64/32b
ABI, 64-bit compat handling and appropriate conversions must be added to
handle the different union/struct sizes caused by this pointer size
mis-match.

Before extending the number of bits in union bpf_attr to accept
capabilities, set the groundwork for handling compat64. When
in_compat64_syscall(), take the compat64 sized bpf_attr and convert it
to what will be the new native offsets.

Inbound conversion is handled upfront to minimise impact on existing
code and reduce overall diff size. After dispatch_bpf the majority of
code can remain unchanged. The cases where conversion back out to
userspace is required are handled in subsequent commits.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>