lib: Refactor do_strnlen_user()
The logic to process aligned words in do_strnlen_user() is fairly
complicated. This patch simplifies it a little bit, notably by
moving it to its own function.

This refactoring is mainly done in preparation to a further patch
that will prevent under- or over-reading the input string, when
doing so is invalid.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>