tcp: Implement compat64 handling for struct tcp_zerocopy_receive
Introduce a compat64 version of the struct tcp_zerocopy_receive
structure. Also, implement helper functions that convert
between the compat64 version and the native version of the struct.

A subsequent patch is going to change the struct tcp_zerocopy_receive
to enable it to support new architectures. On such architectures, the
current struct layout still needs to be supported for compat64 tasks.

Signed-off-by: Harry Ramsey <harry.ramsey@arm.com>
Co-developed-by: Tudor Cretu <tudor.cretu@arm.com>
Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>