tcp: Support capabilities in tcp_zerocopy_receive
The tcp_zerocopy_receive struct is used to create a shared memory region
between userspace and the kernel in which received network packets are
directly copied from the network interface card.

In the PCuABI, a user pointer is a 129-bit capability, so the __u64 type
is not big enough to hold it. Use the __kernel_uintptr_t type instead,
which is big enough on the affected architectures while remaining 64-bit
on others. Additionally, use the special copy routine when copying user
pointers from and to userspace.

An additional check has been added to ensure that the capability has
ownership on the mapping backed by the TCP socket fd to receive network
packets.

Signed-off-by: Harry Ramsey <harry.ramsey@arm.com>
Co-developed-by: Tudor Cretu <tudor.cretu@arm.com>
Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>