lib: Make strnlen_user() PCuABI-friendly
strnlen_user() currently assumes that it is able to read the entire
user string word by word, potentially under- or over-reading.

Unfortunately such assumption is invalid in PCuABI, as the bounds of
the input user (capability) pointer may exactly match that of the
string, preventing any under- or over-read, even by a few bytes.

do_strnlen_user() is therefore amended to check the bounds (base and
limit) of the input pointer. If the bounds do not include the entire
first or last word, the head and/or tail characters are read byte by
byte (find_zero_unaligned()). The rest of the string is still read
using the faster, word-by-word find_zero_aligned().

This change should be a no-op for standard ABIs (retaining the
unconditional word-by-word loop), as the bounds of non-capability
pointers are considered unlimited.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit b57a17d3443ac208dba1248f294403b5d7aa47a5)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>