arm64: vDSO: Provide a purecap vDSO to userspace purecap programs
Map the purecap vDSO into userspace application address space, therefore
providing purecap userspace applications a purecap vDSO to use, and
provide a capability to it in AT_SYSINFO_EHDR by updating
ARCH_DLINFO. Redefine the aarch64 vDSO as the compat vDSO, therefore
allowing both regular arm64 and purecap binaries to use a vDSO under
PCuABI.

Now that the purecap vDSO is mapped in purecap processes, instead of
the standard aarch64 vDSO one, we also need to ensure that
setup_return() uses the right offset for the sigret trampoline.
This is done by including only the appropriate offsets header (which
is possible as signal.c is built twice, once for purecap and once
for compat64).

The bootstrap Morello kselftest is also amended to expect
AT_SYSINFO_EHDR to be a valid capability, not a null one.

--
CHERI-Linux (Hesham): cherry-picked 21ed9c7 and fixed merge conflits:
* Removed VVAR
* Aligned with the current purecap-vdso paths and builds
* Reduced code duplication by passing abi to
__arch_setup_additional_pages.
--

Signed-off-by: Aditya Deshpande <aditya.deshpande@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Hesham Almatary <hesham.almatary@cl.cam.ac.uk>