mm/mincore: Add PCuABI capability constraints
Capability permissions and bounds constraints are added as per the
PCuABI specification for the mincore syscall. mincore() does not
require VMem and only one of the RWX permissions, so the standard
check_user_ptr_owning() interface is not used, and permissions are
verified explicitly.

Also, as mincore() allows the address range to not span whole pages,
the bounds checking is tweaked accordingly.

Signed-off-by: Amit Daniel Kachhap <amitdaniel.kachhap@arm.com>
Co-developed-by: Kevin Brodsky <kevin.brodsky@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>