arm64: morello: Implement user capability accessors inline
We no longer need to have standalone
__morello_{get,put}_user_cap_asm() helpers; we can implement the
access to user capabilities inline, on the model of
__{get,put}_mem_asm(). We cannot directly use the latter though, as
a different constraint is required for the capability to load/store.

There is currently no way to get Clang to output the X register
corresponding to a capability argument, so we also need to add
.L__gpr_num_* aliases for C registers, in order to keep the
EX_DATA_REG() macro happy.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>