mmc: Support capabilities in MMC_IOC_{MULTI_}CMD ioctls
The mmc_ioc_cmd and mmc_ioc_multi_cmd structs are used to hold information
and data about an MMC ioctl.

The PCuABI uses 129-bit capabilities as user pointers, which means that the
__u64 type must be replaced with the __kernel_uintptr_t type, which is
large enough to hold capabilities, yet will remain 64-bit on other
architectures. Additional modifications have been made to the copy routines.

The command value for an ioctl is inferred from the size of the struct that
is provided to it; this patch increases the size of struct mmc_ioc_cmd with
the use of __kernel_uintptr_t, which in turn modifies the calculated value
for the ioctl. In order to prevent this value changing, it is necessary to
hard-code the definition of the MMC_IOC_CMD ioctl value, as compat structs
are not defined in this uAPI header. Because the alignment of struct
mmc_ioc_cmd increases too, struct mmc_ioc_multi_cmd also grows and
MMC_IOC_MULTI_CMD is hard-coded in the same manner.

Signed-off-by: Akram Ahmad <Akram.Ahmad@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>