fs/proc: Update pagemap ioctl handler for PCuABI
Switch the pagemap ioctl handler to the new PCuABI-friendly
prototype, taking user_uintptr_t as third argument, and propagate it
in helper functions. The argument is always a pointer so use
compat_ptr_ioctl as compat handler.

Minimal changes are also made to ensure that the manipulation of
pointers related to the fields of struct pm_scan_arg is valid in
PCuABI. More comprehensive changes will be needed to pass through
capabilities from userspace (at least for the vec field of the
struct).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>