arm64: Kconfig: Use native mmap_rnd_bits in compat64
The ARCH_MMAP_RND_COMPAT_BITS* values are only appropriate for
AArch32. For compat64, we do not need separate values, as the
address space size is the same as in native.

This patch restores the number of ASLR bits to its appropriate value
in compat64 by unselecting HAVE_ARCH_MMAP_RND_COMPAT_BITS, so that
the native values are used in compat64 too.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>