kselftests/arm64: morello: Use O_TMPFILE for temporary files
A few tests currently create temporary files in / to obtain an fd
for testing further syscalls. This is suboptimal as these files may
be used for other purposes, they are not deleted after the tests are
run, and we should generally avoid creating files at the root.

We could create files in /tmp instead, with some naming pattern à la
mkstemp(), but it turns out there is an even simpler solution.
open() can be passed O_TMPFILE to create an unnamed file under a
given directory. The file disappears once the fd is closed, so there
can be no leftover. This approach is used to implement tmpfile() in
glibc, if available.

On the model of tmpfile(), add a new helper tmpfd() in
freestanding.h and make use of it in the tests that need a temporary
fd, replacing the hardcoded files. We try to create the unnamed file
under /tmp if it exists, as /tmp should be writable by any user,
while / may not be writable at all.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit 6765c754ab2cb281c4fbdb439c02f2e70237782e)
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>