arm64: compat: Create null-derived out-of-bounds pointers
In PCuABI, we need compat_ptr() to create a valid capability from an
arbitrary compat pointer, to enable uaccess to the underlying
memory. However, if the compat pointer is clearly an invalid
address, it is preferable to return a null-derived (invalid)
capability. This is especially true when the address is null: we
clearly want to return the null capability in that case.

This should be a non-functional change (uaccess would fail anyway
due to the absence of underlying mapping), but it avoids some
confusion when debugging compat64 handling (and reduces the creation
of capabilities).

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>