riscv: hobgoblin: Xilinx EthernetLite support
This patch add in support for the Xilinx EthernetLite IP on the
GenesysII board. This is a very simple Ethernet controller. It does not
do DMA, so all data is transferred via the CPU. Performance is low,
especially as the CPU on the hobgoblin platform is only running
at 50MHz or so.

The dts file is called a70x-hobgoblin-etherlite.dts as there will be a
version using the Xilinx axi-ethernet which is a much more capable
controller, but requires a license to build. The non ethernet version is
removed from the build to remove clutter.

Signed-off-by: David McKay <david.mckay@codasip.com>