riscv: hobgoblin: Xilinx Axi Ethernet support
This adds support on the hobgoblin platform for the Xilinx Axi Ethernet.
This is a much more capable Ethernet than then EthernetLite. It supports
DMA and TCP checksum offload.

Performance is much improved over the EthernetLite, the downside of
this design is that it does require a license for the Ethernet block.

Added in AXI interconnect clock and removed the spurious clock that was
there.

Signed-off-by: David McKay <david.mckay@codasip.com>