net: emaclite: Use xemaclite_{readl,writel} to access ping/pong buffers
The memory mapped ping/pong buffers were previously being accessed
simply by dereferencing pointers. While the pointers were of uint32_t
type, the compiler is not obliged to perform a 32 bit access, and it
has been seen that clang 18 will change the access to a byte read for
the final copy. In addition the pointer has been obtained by calling
ioremap(), so should never be treated as a normal pointer anyway.
Update the code to call xemaclite_{readl,writel} like any other
register. This also allows the memory barriers to be removed, as
they are implicit in readl/writel.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>