binfmt: Store initial user pointers in bprm in PCuABI
In PCuABI, the initial pointers (program counter, stack pointer,
argv/envp/auxv) are written to registers in the arch-specific
START_THREAD(). argv/envp are also stored in the auxiliary vector
itself.

We are about to narrow down the bounds and permissions of those
pointers, so to avoid duplicating capability operations, make the
pointers directly available to START_THREAD() by storing them in the
bprm struct.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>