arm64: Use bprm->pcuabi to set initial pointers in PCuABI
The bprm struct now contains the full initial user pointers (PCC,
CSP, argv/envp/auxv) as calculated by binfmt_elf, so we can use them
to set pcc, csp and c1-c3 in start_thread() in PCuABI. This also
ensures that the capability metadata of argv/envp is the same in the
auxiliary vector and c1-c2.

To centralise the PCuABI handling in morello.c, x0 and c1-c3 are now
set in morello_thread_start() instead of the separate
init_gp_regs().

start_thread() had to be moved to process.c as it is not possible to
include linux/binfmts.h from asm/processor.h (circular reference),
in order to obtain the definition of struct linux_binprm.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>