kselftests/arm64: morello: Process caprelocs using appropriate root caps
We are about to restrict the bounds of initial capabilities, such as
PCC, and nullify DDC, in line with the PCuABI specification.
Additionally, Clang/LLD still generate caprelocs when linking a
static executable, so we need to ensure that we process them in a
way that is PCuABI-compliant.

Accordingly, this patch switches the caprelocs processing to using
AT_CHERI_EXEC_*_CAP as root capabilities instead of DDC, just like
when processing dynamic relocations.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>