fs/binfmt_elf: Restrict executable/interpreter capabilities
This patch narrows down the bounds and permissions of all initial
capabilities pointing to executable and interpreter mappings in
PCuABI, as per the specification. This includes:

* PCC (the initial capability program counter)
* The root executable/interpreter capabilities
  (AT_CHERI_{EXEC,INTERP}_*_CAP)
* Other auxv entries that point to the executable (AT_ENTRY, AT_PHDR)

The capabilities are currently derived from the root userspace
capability. While the resulting bounds should be correct, they
should ideally be derived from the appropriate owning capability
instead, once reservations are implemented.

AT_BASE remains a simple address in the PCuABI specification, this
will be amended in a subsequent patch.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>