fs/binfmt_elf: Restrict stack capabilities
This patch narrows down the permissions, and bounds if possible, of
all initial capabilities pointing to the stack in PCuABI, as per the
specification. This includes:

* CSP (the capability stack pointer)
* The argv/envp/auxv pointers
* The root stack capability (AT_CHERI_STACK_CAP)
* Other auxv entries that point to the stack (AT_RANDOM, AT_EXECFN,
  AT_PLATFORM, AT_BASE_PLATFORM)

Capabilities giving access to the whole stack (CSP,
AT_CHERI_STACK_CAP) remain unbounded for now, as the stack is itself
unbounded. It will be possible to reduce their bounds once a stack
reservation with a fixed size is introduced, as per the PCuABI
specification.

argv and envp may overlap if a very large number of arguments /
environment variables is provided (see TODO in
set_bprm_stack_caps()). This is however unlikely in practice.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>