kselftests/arm64: morello: Seal reconstructed function pointer
switch_to_restricted() must explicitly reconstruct the given
function pointer, as it needs to remove a permission bit, which
clears the tag due to the capability being sealed. However, it does
not currently reseal the capability after using the BUILD
instruction. This will cause a capability fault on BLRR once
CCTLR_EL0.SBL is set, as per the PCuABI specification.

Fix this by explicitly sealing the new capability before branching
to it.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>