arm64: barrier: Make arch_counter_enforce_ordering() purecap-friendly
The vDSO uses arch_counter_enforce_ordering(), so it needs to work
as intended in purecap (when PCuABI is selected). At the moment it
creates a dummy pointer from the 64-bit value of SP and then
dereferences it. To make this valid in purecap, derive a valid
capability from CSP instead.

This issue went unnoticed due to DDC allowing all X-based
loads/stores, even in purecap. Once DDC is nullified, this patch is
required to avoid a capability fault when
arch_counter_enforce_ordering() is called from the purecap vDSO.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>