arm64: ptrace: Preserve capability in tls_set()
In PCuABI, ptrace(PTRACE_SETREGSET, REGSET_TLS) currently sets the
capability thread pointer (CTPIDR) to a 64-bit value, zeroing any
existing capability. This is unlikely to be the desired behaviour,
and is inconsistent with the register merging principle, which the
other ptrace operations obey. For instance, setting a new value for
SP using ptrace(PTRACE_SETREGSET, REGSET_GPR) does not zero CSP, but
instead sets its address to the provided value.

Unlike other capability registers like CSP, CTPIDR is stored only as
a capability (in tp_value). We therefore need to perform the X/C
register merging manually. The corresponding function from morello.c
is exposed in asm/morello.h for that purpose.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>