mmc: ioctl: Remove hard-coded MMC_IOC_{MULTI_}CMD values
The "mmc: Support capabilities in MMC_IOC_{MULTI_}CMD ioctls" patch
previously hard-coded the ioctl command value in order to maintain
compatibility with non-PCuABI ioctl handlers such as those found in
block.c. This deviates from the standard approach which is to always
define the ioctl command value in terms of the struct size (which may
change), instead providing any fixes in the kernel via compat ioctl
handlers.

This patch therefore aligns the MMC ioctl handlers in block.c with
the standard approach. Undoing the hard-coding in the header leaves
the ioctl command value unchanged in all situations except in PCuABI
where the structs sizes, and hence the ioctl command values, are
now larger (the ioctl value remains unchanged in compat64 as well).
The additional comments that were added in the aforementioned patch
regarding the changing size of the structs is also removed in this
patch.

Signed-off-by: Akram Ahmad <Akram.Ahmad@arm.com>
Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
(cherry picked from commit de68fdb44b6096016744d863c92dc83fe6c2be70)
[ Christian A. Ehrhardt: Conflict resolution ]
Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>