drm: Fix copy to/from user so that caps transport in the region
DRM makes copies of ioctl payload data to/from userspace and some of
this can also have pointers to other memory (or capabilities for
morello) and thus it needs to assume the worst and copy while
retaining capabilities. This also requires making sure the stack
buffer for small ioctls (128 bytes or less) is aligned to the maximum
pointer or capability size.

Signed-off-by: Carsten Haitzler <Carsten.Haitzler@arm.com>