fs/exec: Create a stack reservation in PCuABI
The stack needs special handling in PCuABI to ensure that its
underlying reservation is appropriately sized. A new sysctl
"cheri.max_stack_size" is introduced to that effect, as per the
PCuABI specification, specifying the reservation size (before
alignment) unless the hard limit of RLIMIT_STACK is lower (unlimited
by default).

The default value for cheri.max_stack_size (128 MB) should be plenty
for any application. The minimum value (256 KB) was chosen to leave
a safe amount of space over the initial stack mapping expansion (128
KB, see setup_arg_pages()). The maximum value is fairly arbitrary,
but has the advantage of fitting in an int.

acct_stack_growth() is also modified to prevent a mapping with
VM_GROWSDOWN from growing beyond its reservation (for this reason
the reservation details are set before calling
expand_stack_locked()). This is only used for the main stack, as
PROT_GROWSDOWN is not supported in PCuABI.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>