Documentation: cheri: Update PCuABI status
Our implementation of PCuABI is now fairly close to the full
specification, and the transitional PCuABI specification is no
longer relevant.

Update the documentation to remove references to the transitional
ABI, and list the remaining limitations w.r.t. the full ABI.

The Morello-specific limitations are also updated to be more
representative of the current status. Those related to mmap-like
syscalls in hybrid, and ptrace in general, are removed as they are
the results of design decisions that have not expected to change.

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>