riscv: hobgoblin: Add A730 MP4 SMP for proFPGA
This patch adds in support for an MP4 SMP cluster on the proFPGA board.
The proFPGA board is only supported in the MP4 configuration to reduce
the number of configurations.

Unfortunately, due to the limitations of the Xilinx interconnect we have
to place the main DDR memory outside of the 4G 32 bit memory area. This
means that the proFPGA board has no memory at the 0x80000000 address,
resulting in bootloaders and the like having to be built at a different
address. It also means that it is impossible to use any 32 bit only DMA
device.

It should be possible to overcome this by constructing and alias of the
DDR memory at the 0x80000000 address, but this is still experimental.

It would be possible to just have the MP4 CPU description and disable
the CPUs that are not present on particular boards. However, doing it
with distinct files results in a cleaner dtb file and is probably less
confusing to the end user.

One thing to note that was discovered while doing this work is that the
order of the DT entries does matter. The reference to the PLIC
interrupts was originally done via a reference back. This meant that the
PLIC entry appeared before the interrupt controller. On 5.x Linux
kernels this meant that the kernel would not boot as it was looking for
an interrupt controller to be present when parsing the PLIC entry. Later
kernels do not have this problem.

As a result of this, I have been very conservative and tried to make the
order as traditional as possible to avoid any nasty surprises.

Signed-off-by: David McKay <david.mckay@codasip.com>