cheri: Provide an atomic pointer type similar to atomic_long_t
For CHERI the atomic_long_t is not sufficient to do atomic
operations on pointers. Provide a suitable type and associated
atomic operations for pointers.

For targets where the size of a pointer matches the size of
an unsigned long the implementation falls back to the existing
atomic_long_t.

We use CONFIG_CHERI_KERNEL to determine if a special atomic_ptr_t
type is needed. Using __SIZEOF_POINTER__ > __SIZEOF_LONG__ fails
at least for arm64 with purecap UABI when compiling the VDSO.

An implementation for RISCV with CHERI will be provided in a
later change.