riscv: hobgoblin: Add second Ethernet controller
The VCU118 has to use the FMC-Lite add in board in order to get access
to an SD card. This board also has an Ethernet Phy on it is well, so
this patch adds support for it as the second Ethernet. It is the same
AXI enmac controller.

The FMC-Lite uses a MicroSemi VSC8541 phy in RGMII mode. On the VCU118
it is necessary to add in a 2ns delay on the RX side in order for the
receive side to work correctly.

Signed-off-by: David McKay <david.mckay@codasip.com>