cheri: More uintptr_t vs. unsigned long fixes
There is one issue with lists that store a pointer and additional
flags in a void * or uintptr_t field. If the low bits of the
(aligned) pointer are used, we are ok. However, if the pointer is
shifted somehow the address will go out of bounds and the pointer
cannot be recovered. This is marked with FIXCHERI for now.