cheri: Fix bounds violation in strncpy_from_user
Callers of strncpy_from_user gracefully handle page faults and
work with the amount of data the actually was available. This is
neccessary whenever the API does not provide an explicit length
for a string.

However, the algorithm to do this assumes that a read access beyond
the end of the string is ok. For the CHERI case avoid a CHERI fault
by limiting the maximum read size to the capability bounds.