cheri: riscv: Somewhat support RISCV_ISA_ZBB for cheri
Make the kernel compile if RISCV_ISA_ZBB is enabled. Most of the
optimized code is not ready for CHERI and unused in a CHERI kernel
as the implementations assume that read access outside the string
(or mem region) is ok as long as some of the algined access is
within the string.