cheri: Refactor Kconfig logic for CHERI
Allow the architecture to declare that it only supports certain
combinations of purecap UABI and cheri kernel.

For ARM declare that it only support purecap UABI or no cheri.
For RISCV declare that purecap UABI is possible iff the kernel
is cheri.

The riscv only use CONFIG_CHERI_KERNEL in ARCH dependant code.