cheri: Return properly bounded pointers from gen_alloc_*()
Adjust the generic allocator in lib/genalloc.c and its callers to
return properly bounded pointers. The allocator is used by some
callers to manage a range of things that is not virtual memory and
thus does not have to be a valid pointer. Check what the caller
provides and only set bounds if we are in fact dealing with pointers.