cheri: codasip: Enable zicbom and zicboz extensions
The zicbom and zicboz extensions are supported by the a730
hardware and thus announced in the device tree provided ISA
string. However, qemu does not support these extension which
leads to confusion when they are announced in the device tree.

Disabling the option in either the kernel or the device tree
is not an option as the resulting kernel would not work with
DMA on real hardware.

As a solution/workaround:
- Enable the extensions in our kernel config.
- Make things compile with CHERI
- Probe for the presence of the extension and remove them
  from the ISA string  if the "hardware" does not support them.
- If the zicbom extension is announced by the ISA string but not
  supported in hardware assume that we are running on qemu and
  no cache flushing is needed even for non-coherent DMA devices.
  This option is configurable and the configuration option is
  enabled for our config.

This will allow us, to run the same kernel on qemu and on real a730
hardware.