riscv: cheri: Fix trap vector in probing code
The probing code for various spec versions and other features
sets the trap vector and then tries an instruction. If the
instruction works as expected the feature is flag as supported.
Otherwise recovery code at the trap vector will mark the feature
as unsupported.

However, the trap vector code was not properly align. As per the
riscv spec the trap vector code must start at a 4-byte aligned
address with the lower two bits clear. Align the trap vector
address accordingly.