cheri/uapi: Improve implementation of __ALIGN_KERNEL macro
This marcro is supposed to align an arbitrary value (pointer or
otherwise). For pointers type in C the type of the mask must
be a ptraddr_t (aka unsigned long) and not a full uintptr_t.
Otherwise the source of provenance is unclear.

Use _Generic() to realize this. The previous implementation used
macros that are not available in UAPI headers. The new version is
at least compatible with C11 enabled userland.