riscv: Make CHERI ucontext the same size as non-CHERI
Make sure that the user context in a signal frame (everything
before the saved register state) has the same size on CHERI
and non-CHERI systems. This will allow unwinders to correctly
locate the signal frame.

To do this we eat into space reserved for an (unlikely) increase
of sigset_t. While sigset_t might have to increase in the future
it is unlikely that it will grow by more than a few words.