riscv: Remove unwind annotations from sigreturn
Remove the unwind annotations from sigreturn. Either the information
should not be there or we must provide enough information to correctly
recover the caller's stack pointer and pcc.

Otherwise unwinders and gcc will be confused and even crash on CHERI
e.g. if a non-NULL but otherwise invalid PC is encountered.