cheri: riscv: ptrace: Provide tag information in GPR register set
When the guest gets the GPR register set via ptrace(),
the register tags will be cleared when copying to user space.
Add a pseudo register at the end of the register set that contains
the original values of the tags to allow the debugger to display
tag information for registers.

For the corresponding "set" operation try to create capabilities
in the tracee's registers by taking provenance from the old
register value. This covers the most common cases where the
debugger writes to registers and avoids security concerns that
arise from being able to create arbitrary tagged capabilities
in the tracee.