cheri: riscv: Implement PTRACE_PEEKCAP
All the ptrace operations will copy capability data from the
ptraced task without tags. The tagged capabilities would have
have no meaning in the ptracer and the tracer could abuse this
to mint capabilities that it should not have.

While the debugger does not need tagged capabilities it does want
to know about the tag values and display them during debugging.

To this end introduce the PTRACE_PEEKCAP call that will retrieve a
capability size guest region and its tag.