cheri: mm: Refactor reserv management
Currently, the way memory reserves are managed has the hidden assumption
that an mmaped range is reasonably aligned within the minimal
representable range that is surrounding it. E.g. when creating a
reservation but also when checking reservations there is the assumption
that for a given range [addr,addr+len] is contained within the range
that is optained when rounding addr down to a representable address
and length up to a representable length. However, if the start address
is close to the end of an aligned chunk with the size of the
representable length the rounding down of the address might cause
the resulting length to not cover the end of the original range.

This happens at least when unmapped space is allocated through the
huge page code for file mappings. In this case the odd alignment wrt.
a huge page to make file offset and huge page offset match might
interfere with the attempt of the CHERI code to align the range on
a representable address boundary.

Fix fix this track two ranges in a reservation:
- The inner range the _only_ covers the range that the actual
  reservations is for (i.e. what the user requested). This range
  has no alignment constraints (above PAGE_SIZE alignment).
- The outer range which is what the boundaries of an owning
  capability will be.

The inner range is used whenever the check is for inclusion,
i.e. a range is only included (completely) in a reservation if
it is within the inner bounds. The outer range is used for
conflict checking, i.e. a different reservation's outer range
must not overlap with the outer range of this existing reservation.