usb: Fix unaligned allocations in hcd_alloc_coherent
hcd_alloc_coherent allocates data from a pool and subsequent
allocations might not be aligned if a previous allocations
requested an unaligned number of bytes.

Make sure that all allocations are at least pointer size
aligned. This allows us to store capabilities in the
allocated chunks.

This is necessary because we store a pointer right after the
allocated chunk that can be used to recover the buffer.