cheri: riscv: Fix definition of CHERI_PERMS_ROOTCAP
The current definition of CHERI_PERMS_ROOTCAP pre-dates the
introduction of Zcherilevels into the riscv-cheri spec.

Now that levels are there, fall back to the default definition
from include/linux/cheri.h. Make sure that we handle the level
correctly if the hardware does not support levels.

As a result both the root kernel and the root user capability
will have all the levels permissions enabled. The kernel does
not make any use of Zcherilevels for now but userland could.

The main purpose (prevent kernel capabilities from escaping to
userland is not really an issue due to the address space split.
All in-kernel capabilities are restricted to the upper half of
the address space that is not accessible to userland.