cheri: bpf: Fix missing tag on a call in BPF
BPF calls jump to some function in kernel code . The function
is given by an relative offset to __bpf_call_base which itself
is a dummy function. However, doing arithmetic with the pointer
to the dummy function fails as the function pointer is sealed.

Unseal the function pointer with cheri_make_kernel_code_cap().

This removes the type information from the pointer thus cast
the result to the correct function type as appropriate.