cheri: mmap: Fix reservation system some more
Unfortunately, it was still possible that a new unmapped area
was allocated completely within the existing reservation of
another VMA. Refactor the VMA gap calculation and the allocation
of unmapped areas again to really make sure that this does not
happen.

NOTE:
Reservations can be (significantly) larger than the actual mapped area.
Possible reasons are:
- User space maps a larger area and later unmaps parts of it.
- The area is large bug unaligned. In this case the reservation
  must be aligned to a representable range.