cheri: futex: Fix futex hashing
When sleeping and waking up a futex information about the
futex is hashed to and the futex context is looked up based
on this hash in a global hash table. Depending on the futex
type (shared vs. private) different items are used to calculate
the hash.

The function get_futex_key() collects the information and stores
it in a futex_key union. This is a union type that has different
members for different futex types.

The hash is calculated using a variant of the union that is supposed
to be the common denominator of the actual variants and the hash
function assumes that the offsets between the variants match.

Due to the increased size of a pointer in the CHERI case, this
assumption is no longer true and hashing includes data that can
vary between syscall invocations.

To fix this make the union elements match again and add static
assertions to ensure that this remains true in the future.