usb-storage: fix a sparse warning
Fix this sparse warning

drivers/usb/storage/usb.c:593:22: warning: non size-preserving capability to integer cast

id->driver_info is a uintptr_t that contains a set of flags. We can
downgrade the capability to a numeric value.

The next line

usb_stor_adjust_quirks(us->pusb_dev, &us->fflags);

takes the address of the converted flags. This is ok, a proper capability
is created for the second parameter.