wifi: mac80211: modify check of ieee80211_tx_info alignment
struct ieee80211_tx_info looks like this

struct ieee80211_tx_info {
   ... two u32 variables

   union {
      ...
      struct {
         struct ieee80211_tx_rate rates[IEEE80211_TX_MAX_RATES];
         ...
      } status;
   };

The cheri compiler aligns the inner struct status to capability size.
Update the assertion in the code to match this alignment.

struct ieee80211_tx_info is a kernel internal data structure to pass info
between a wifi driver and the mac80211 layer. The difference in alignment
between cheri and non-cheri should not be a problem since the struct is
not passed to the hardware.