wext: adjust the length for packed events
The callers of iwe_stream_add_event_check pass an event length that
includes padding after len and cmd. In other words: They all pass
IW_EV_LCP_LEN + the length of their payload.

Remove IW_EV_LCP_LEN in iwe_stream_event_len_adjust and replace it with
the lcp length we need for the packed event.

This patch does not change the values if CONFIG_COMPAT is defined.

All credit for this goes to Christian.