cheri: Make net/ compile.
Make net/ compile for CHERI, add casts etc.

Disable CONFIG_RDS for the time being. The UAPI issues related
to that need more in-depth analysis.

Add a hack to the xt_counters data structure. It is either used
as two u64 or as one pointer. Thus we can fit the pointer in there.