cheri: acpi: Disable packing for now
The ABI/API for CHERI will need (a lot of) additional thought.
For this reason CONFIG_ACPI is currently made incompatible with
CONFIG_CHERI_KERNEL.

However, some headers are still included and cause warnings due to
the underaligned capabilities in ACPI structures. Disable packing
on these structures for now. This will basically render the structs
incompatible with what hardware provides. Thus add a comment to
make that clearly visible and only doe it for CONFIG_CHERI_KERNEL.