cheri: Fix bounds for reserved pages in __va()
The size of the page allocation in the buddy allocator is used
to calculate the bounds of a capability retrieved from __va()
and related function like page_to_virt() and phys_to_virt().

However, this fails for pages that are not managed by the buddy
allocator but are "stolen" from it by memblock_alloc() during
boot.

Add a mechanism to determine if these stolen regions are eligible
for virt_to_phys() and if so store the length of the block. Due
to the limited amount of space in the struct page only these bounds
will always start at the page that contains the address that is
passed to __va(). They will not allow access to the parts of the
memblock before this.