PCI: xilinx: Optionally suppress correctable errors
If your PCIe links are less than stellar, you will get a message printed
every time a correctable error is seen. This can be expensive if there
are a fair number being generated.

This provides an option to disable them, so the interrupt will not be
taken. It would also make sense to rate limit the messages if they are
enabled as well.

Signed-off-by: David McKay <david.mckay@codasip.com>