riscv: increase kernel stack size for cheri
Increase the stack size of kernel threads on cheri systems.

Overflowing such a stack causes a bounds violation of the stack pointer
capability. This triggers a cheri exception which tries to save more data
on the stack and triggers yet another exception...

Our testing showed that for the current kernel stack size, it's
relatively simple to fill up the stack without any fundamental
misbehaviour of the system. We downloaded large files over wifi with
lockdep enabled. The tx and rx worker threads are interrupted by usb
interrupts, these interrupts use the stack of the thread that is
interrupted. The lockdep routines store much data on the stack.

With the increased kernel stack size, the wifi download tests worked
reliably without any deadlocks.