cheri: Silence linker warnings for zero sized structures
The CHERI linker does not like zero sized symbols in object files
because it cannot determine if the object size is just missing or
actually zero. Thus a warning is emitted if such an object is
encountered during linken.

Avoid these warnings by adding an additional byte the size of the
generated objects.