sparsemem: fix __pfn_to_page for cheri, no vmemmap
Update the __pfn_to_page macro to work on cheri systems that use
sparsemem without vmemmap.

Without the global vmemmap array, we have no suitable capability from
which to build the struct page pointer that __pfn_to_page returns. The
section's section_mem_map points to a location far outside of the actual
struct page array. (It's set up such that section_mem_map + minimum pfn
of the section points to the start of the array.)

We could add a capability to each struct mem_section. The struct must be
power of 2 sized. Adding an element would effectively double its size.

The easiest solution is to keep the non-cheri address calculation and
build a capability with the resulting address. The bounds have to cover
the whole struct page array. There's lots of code that moves a struct
page pointer within the array (e.g. "buddy pages").