cheri: Apply automatic fixes from the FixCapAsArraySubscript rule
Automatically wrap array indices that have capability type
in a call to __c_ua(). Otherwise the compiler will crash
if -fsanitize=array-bounds is enabled.

We can decide if we want to keep the patch or drop it once
the compiler is fixed.

Ignore changes in arch/riscv/mm/init.c which has interesting
uses of uintptr_t that will be cleaned up later.