cheri: Fix suspected upstream bug in sound/soc/codecs/ab8500-codec.c
The CHERI clang-tidy checker complains about incompatible pointer
target types in the target of a cast in ab8500_codec_probe().

Apparently, the code is in fact wrong: The "private_value" field
is a pointer to a "struct filter_control" encoded as a uintptr_t.
This matches the code in the initialization. However, when using
the "private_value" its address and not its contents are cast to
"struct filter_control".

Given that the code is there since 2012 it must be either mostly
unused or I am missing something.

Thus keep this as a separate commit for the time being.