riscv: increase overflow stack size for cheri
If vmap stack is enabled, there's guard pages around the stacks of kernel
threads. When a stack overflow is detected, the exception handler uses
the per-cpu overflow stack to avoid an endless loop.

Increase the overflow stack size from 4k to 8k for cheri kernels. On a
cheri system, 4k are enough to save the register file, but not enough for
initiating a reboot after a kernel panic.