cheri: Fix xarray handling in drivers/cxl/core/
Use plain addresses instead of pointers to index an xarray.

In one case the upstream code tries to convert an xarray index
back to a pointer which does not work with CHERI. For now, we
don't want to allow this in xarray code because it would complicate
things if some users want the feature and others don't.

Thus work around the issue in the CXL core itself by storing
the pointer in the entry. The xarray index is still just an
address.

We may have to revisit this if more xarray users like this pop up
in the future.