riscv: cheri: make asm_per_cpu work on cheri kernels
Make the asm_per_cpu macro work when cheri is enabled in the
kernel config.

Use lgc to load the proper capabilities for __per_cpu_offset and the sym
parameter. llc CREG(\dst), \sym would create a capability with ppc's
metadata and sym's address. Such a capability would not never grant write
permission (since ppc does not grant W) and would not be usable as a
stack pointer.

Build the return capability in the same way as the C macro this_cpu_ptr.
Make the return capability's base address identical to the capability's
address to get away with a single tmp register. This is simpler than the
C macro, but should not be a problem in practice.

Set the other components of the return capability in a way that works
with a single tmp register. After setting the address and before setting
the bounds, the address might be unrepresentable and the tag might be
cleared. Set the tag again at the end of the macro when all fields
contain their final values.