cheri: riscv: Provide provenance when patching alternatives
When patching an alternative the riscv code uses an offsets
from the location of the alternative to point to the location
to patch and the replacement code. This offset neither has the
correct bounds nor the necessary permissions to do the patching.

Menually derive the alternative pointer for these cases.