cheri: Fix percpu reference counters for CHERI
The counter can be either global or (on demand) a percpu pointer.
In the first case the counter is just a counter but in the second
case it is in fact a percpu pointer. Thus the data type must allow
for storage of a uintptr_t and not just an unsigned long.