cheri: fs: notify: Fix uintptr_t usage in fs/notify/
Use uintptr_t for values that can contain a pointer in fs/notify/
and cast as necessary.

Some fanotify events contain an embedded buffer that can hold
small payload data but must be replaced with an external buffer
if the payload data is too large. In the latter case the inline
buffer contains a pointer to the external data. Thus make sure
that the inline buffer can contain at least the header and one
aligned pointer even for CHERI.