fs: erofs: Fix CHERI issues in fs/erofs/
The ->private field of a folio is used by erofs to store an
atomic_t directly in the private field. This results in CHERI
related warnings because a pointer of type "uintptr_t *" (the type
of "&folio->private") is cast to a pointer of type "atomic_t *" which
has a different pointer target size.

Generally, such a cast is a problem because folio->private (despite
being a uintptr_t) will be accessed as a plain address. However, in
this case it is ok. Add an intermediate cast to "void *" to silence
the warning.