cheri: Use uintptr_t sk_buff
For CHERI several fields in struct sk_buff must be able to hold full
capabilities. Change those to uintptr_t and adjust users. The field
"dev_scratch" can not hold a pointer. Instead the storage of the
pointer @dev pointer is used to store non-pointer data that fits into
an unsigned long.