cheri: net: Fix xt_counters for CHERI
The structure is defined in the UAPI headers and it is used to
communicate counter values to userland.

However, inside the kernel the same data structure is used but
it may be re-interpreted as a percpu pointer to a list of counters.
Without CHERI this goes unnoticed by the UAPI.

For CHERI we have to fit a capability into that structure. Thus
make the structure a union between to u64 and a capability. This
is ok as the memory either contains tow u64 value or one pointer.
As userspace should only look at the verion with two u64 counters
this does not break the ABI.

However, the kernel hack should be removed from UAPI headers.