cheri: net: Fix xt_counters for CHERI
The xt_counters structure is defined in the UAPI headers and it
is used to communicate counter values to userland.

However, inside the kernel the same data structure is used but
it may be re-interpreted as a percpu pointer to a list of counters.
Without CHERI this goes unnoticed by the UAPI.

For CHERI we have to fit a capability into that structure. Thus
make the structure a union between to u64 and a pointer. This
is ok as the memory either contains two u64 values or one pointer.
As userspace should only look at the version with two u64 counters
this does not break the ABI.

Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>