cheri: net: Use 16 bit for nf_ct_ext offsets
The total size of all possible extensions exceeds 256 bytes on CHERI
systems. Increase the size of the offset array entries to 16 bit.

Adjust the data alignment for CHERI. Conveniently, the allocation of
members aligns each element to the same alignment as struct nf_ct_ext.

Apply bounds to the individual extensions.