cheri: net: sched: Fix pointer value handling in em_meta.c
The "value" member of "struct meta_obj" must be able to hold a
capability in some cases. However, most of the time it will hold a
plain value.

To avoid changing too much code replace it with a anon union that
has a plain unsigned long member and a new member that can hold
a capability.

Use the capability member where necessary.