cheri: net: bluetooth: Handle pointers in the l2cap_core.c
The API used to set and get options uses an option length and
an option value. If the length is 1, 2 or 4 the option value
is an immediate value of that size. Otherwise the option value
is a pointer that points to the actual data. This data is then
copied to/from the option buffer.

Make the option value a uintptr_t and adjust callers. Most of
the callers that use an immediate value do not need any annotations
as the uintptr_t value is assigned to/from a short integer that
can never be a pointer.