cheri: Allow and correctly handle uintptr_t in array_index_nospec
With ioctl arguments it is possible that a uintptr_t ioctl argument
is passed to array_index_nosepc(). Handle this case gracefully and
return something of the same type.

The somehwhat more premission build time check on the size of the
input value is harmless because we add a case of that value to
unsigned long. Thus if it is not a pointer or uintptr_t this will
fail at compile time.