cheri: wireless: intel: Fix uintptr_t use
- Align structures that contain pointers.
- Add uintptr_t casts.

There is one cast that looks like an upstream bug:
"data->replay_ctr" is of type "u8 *" which suggests that it
already _points_ to the replay counter. However, when passing
the argument to be64_to_cpup the address of data->replay_ctr
is taken which apparently results in the _pointer_ being endian
converted and not the data it points to.