ring-buffer: cheri: tighten data pointer's bounds
Data events in the ring buffer store the length either in type_len or in
array[0]. The length field in type_len is in units of RB_ALIGNMENT, the
length field in array[0] includes the size of the length field itself.

Use this length to calculate the bounds length in bytes when we create a
capability that points to event data.

Fabricate the event data capability based on the capability of a struct
ring_buffer_event's array, there's no need to use the global data
capability.

Make sure that bounds update does not try to increase the bounds, this
would clear the tag.